Combine and get rid of duplicate shared messaging types
