+++ frontend/src/citizen-frontend/messages/MessageTypeChip.tsx
@@ -1,8 +1,8 @@
+import { MessageType } from 'lib-common/api-types/messaging/message'
 import { StaticChip } from 'lib-components/atoms/Chip'
 import { accentColors } from 'lib-customizations/common'
 import React from 'react'
 import styled from 'styled-components'
-import { MessageType } from './types'
 
 // TODO is the 20px line-height in StaticChip unintentional?
 const Chip = styled(StaticChip)`
