+++ frontend/src/employee-frontend/components/messages/MessageTypeChip.tsx
@@ -2,12 +2,12 @@
 //
 // SPDX-License-Identifier: LGPL-2.1-or-later
 
+import { MessageType } from 'lib-common/api-types/messaging/message'
 import { StaticChip } from 'lib-components/atoms/Chip'
 import { accentColors } from 'lib-customizations/common'
 import React from 'react'
 import styled from 'styled-components'
 import { useTranslation } from '../../state/i18n'
-import { MessageType } from './types'
 
 // TODO is the 20px line-height in StaticChip unintentional?
 const Chip = styled(StaticChip)`
